+++ esp-hal-embassy/src/time_driver.rs
@@ -6,12 +6,12 @@ use esp_hal::{
     interrupt::{InterruptHandler, Priority},
     prelude::*,
     time::now,
-    timer::{ErasedTimer, OneShotTimer},
+    timer::{AnyTimer, OneShotTimer},
 };
 
 pub const MAX_SUPPORTED_ALARM_COUNT: usize = 7;
 
-pub type Timer = OneShotTimer<'static, ErasedTimer>;
+pub type Timer = OneShotTimer<'static, AnyTimer>;
 
 static TIMERS: Mutex<RefCell<Option<&'static mut [Timer]>>> = Mutex::new(RefCell::new(None));