+++ examples/src/bin/embassy_multicore.rs
@@ -22,7 +22,7 @@ use esp_hal::{
     cpu_control::{CpuControl, Stack},
     get_core,
     gpio::{Io, Level, Output, Pin},
-    timer::{timg::TimerGroup, ErasedTimer},
+    timer::{timg::TimerGroup, AnyTimer},
 };
 use esp_hal_embassy::Executor;
 use esp_println::println;
@@ -56,8 +56,8 @@ async fn main(_spawner: Spawner) {
     let io = Io::new(peripherals.GPIO, peripherals.IO_MUX);
 
     let timg0 = TimerGroup::new(peripherals.TIMG0);
-    let timer0: ErasedTimer = timg0.timer0.into();
-    let timer1: ErasedTimer = timg0.timer1.into();
+    let timer0: AnyTimer = timg0.timer0.into();
+    let timer1: AnyTimer = timg0.timer1.into();
     esp_hal_embassy::init([timer0, timer1]);
 
     let mut cpu_control = CpuControl::new(peripherals.CPU_CTRL);